+++ hs.js
@@ -1,6 +1,8 @@
 var core = require('./lib/core')
 var base = require('./lib/base')
+var list = require('./lib/list')
 
 exports.core = core
 exports.base = base
+exports.list = list